[major] reorg model config
