+++ lmquant/llm/quant/__init__.py
@@ -1,9 +1,8 @@
 # -*- coding: utf-8 -*-
 
-
 from .activation import quantize_llm_activations
 from .bias import correct_llm_bias
-from .config import LlmQuantConfig
+from .config import LlmQuantCachePath, LlmQuantConfig
 from .reorder import reorder_llm
 from .rotate import rotate_llm
 from .smooth import smooth_llm
