+++ lmquant/quant/quantizer/config/__init__.py
@@ -2,4 +2,5 @@
 
 from .base import QuantizerConfig
 from .kernel import QuantizerKernelConfig
+from .module import ModuleQuantizerConfig
 from .tensor import ActivationQuantizerConfig, TensorQuantizerConfig, WeightQuantizerConfig
